backend/requirements.txt
fastapi
uvicorn
google-genai
python-dotenv
pydicom
pillow
supabase
pydantic
langgraph
python-multipart
numpy